Extract board adapter state and logger emission
